refactor: add function module
