+++ src/App.jsx
@@ -1,7 +1,6 @@
 import OrdersContext from './components/context/OrdersContext'
 import { useState } from 'react'
 import MainTabs from './components/Tabs/MainTabs'
-// import TestCont from './components/TestCont'
 import './App.css'
 
 function App() {
@@ -11,7 +10,6 @@ function App() {
     <OrdersContext.Provider value={{ orders, setOrders }}>
       <div className="App">
         <MainTabs />
-        {/* <TestCont /> */}
       </div>
     </OrdersContext.Provider>
   )
